requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
spacy==3.7.5
langdetect==1.0.9
fpdf2==2.8.1
anthropic==0.40.0
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-multipart==0.0.12
